adding area while using the areas form
