+++ client/src/components/AreasOfFocus/index.js
@@ -1,13 +1,57 @@
+/* eslint-disable react/prop-types */
 /* eslint-disable react/jsx-key */
 /* eslint-disable jsx-a11y/label-has-for */
-import React from "react";
+import React, { useState } from "react";
 import "./styles.css";
-import FormAreasOfFocus from "../FormAreasOfFocus/index.js";
+import AreasOfFocusForm from "../AreasOfFocusForm/index.js";
 
 
 const AreasOfFocus = ({ student }) => {
 
-	const { areasOfFocus } = student;
+	const [areasOfFocus, updateAreaOfFocus] = useState(student.areasOfFocus);
+
+	const [area, setArea] = useState({
+		message: "",
+		level: "",
+	});
+
+	const options = {
+		method: "PUT",
+		headers: { "Content-type": "application/json" },
+		body: JSON.stringify(area),
+	};
+
+	const postAreaOfFocus = () => {
+		fetch(`/api/students/${student._id}`, options)
+			.then((res) => res.json())
+			.catch((error) => console.log(error));
+	};
+
+	const addTheLevelToAreas = () => {
+		area.level === "To work on"
+			? updateAreaOfFocus({
+				...areasOfFocus,
+				toWorkOn: [...areasOfFocus.toWorkOn, area.message],
+			})
+			: area.level === "Okay at"
+				? updateAreaOfFocus({
+					...areasOfFocus,
+					okayAt: [...areasOfFocus.okayAt, area.message],
+				})
+				: area.level === "Good at"
+					? updateAreaOfFocus({
+						...areasOfFocus,
+						goodAt: [...areasOfFocus.goodAt, area.message],
+					})
+					: null;
+	};
+
+	const handleSubmit = (e) => {
+		e.preventDefault();
+		addTheLevelToAreas();
+		postAreaOfFocus();
+		e.target.reset();
+	};
 
 	return areasOfFocus ? (
 		<div className="areas-of-focus-section">
@@ -20,10 +64,9 @@ const AreasOfFocus = ({ student }) => {
 				</h3>
 				<div className="area-text-section-red">
 					{areasOfFocus.toWorkOn.map((item, index) => (
-						<>
+						<div key={index}>
 							<button
 								className={ "areas-button-red key-button"}
-								key={index}
 							>
 								{item}
 							</button>
@@ -35,7 +78,7 @@ const AreasOfFocus = ({ student }) => {
 							>
 								X
 							</button>
-						</>
+						</div>
 					))}
 				</div>
 			</div>
@@ -43,10 +86,9 @@ const AreasOfFocus = ({ student }) => {
 				<h3 className="subtitle-text">Okay at...</h3>
 				<div className="area-text-section-yellow">
 					{areasOfFocus.okayAt.map((item, index) => (
-						<>
+						<div key={index}>
 							<button
 								className={"areas-button-yellow key-button"}
-								key={index}
 							>
 								{item}
 							</button>
@@ -58,7 +100,7 @@ const AreasOfFocus = ({ student }) => {
 							>
 								X
 							</button>
-						</>
+						</div>
 					))}
 				</div>
 			</div>
@@ -66,7 +108,7 @@ const AreasOfFocus = ({ student }) => {
 				<h3 className="subtitle-text">Good at...</h3>
 				<div className="area-text-section-green">
 					{areasOfFocus.goodAt.map((item, index)=> (
-						<>
+						<div key={index}>
 							<button
 								className={"areas-button-green key-button" }
 								key={index}
@@ -81,11 +123,11 @@ const AreasOfFocus = ({ student }) => {
 							>
 								X
 							</button>
-						</>
+						</div>
 					))}
 				</div>
 			</div>
-			<FormAreasOfFocus />
+			<AreasOfFocusForm student={student} area={area} setArea={setArea} handleSubmit={handleSubmit} />
 		</div>
 	) : null;
 };
